drivers: spi_ll_stm32: fix TX or RX only DMA transfers
When only a RX or only a TX transfer was requested the DMA never
finished.

For the RX only cause the size on the transfer was calculated by
taking the TX buffer length (0), this caused problems.

For the TX only transfer the RX buffer was set to NULL, this caused
the DMA to acctually writing data to the adress 0x00000000. By using
the dummy destination buffer it now only writes to valid memory.

Signed-off-by: Erwin Rol <erwin@erwinrol.com>